% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_odds_ratio.R
\name{get_odds_ratio}
\alias{get_odds_ratio}
\title{Get posterior distribution of odds ratio}
\usage{
get_odds_ratio(x.star, x.base, beta_posterior, niter, nburn, K, refK)
}
\arguments{
\item{x.star}{x.star matrix for odds ratio}

\item{x.base}{x.base matrix for odds ratio}

\item{beta_posterior}{(list) posterior distribution of betas from model fit}

\item{niter}{total number of MCMC iterations for inference}

\item{nburn}{number of burn-in iterations}

\item{K}{number of outcome categories}

\item{refK}{index of reference outcome category}
}
\value{

}
\description{
Get posterior distribution of odds ratio
}
