% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_sim_results.R
\name{get_sim_results}
\alias{get_sim_results}
\title{Get parameter estimation and imputation results from a simulation}
\usage{
get_sim_results(fit, ycomplete = NULL, beta_true, gamma_true = NULL)
}
\arguments{
\item{fit}{object of type "pgmultinom"}

\item{ycomplete}{(optional) complete outcome data}

\item{beta_true}{list of true exposure regression coefficients for each category k = 1,...,K-1}

\item{gamma_true}{(optional) list of true covariate regression coefficients for each category k = 1,...,K-1}
}
\value{
a list with components
\itemize{
   \item rmse_beta: RMSE for beta 
   \item bias_beta: bias for beta 
   \item rmse_gamma: RMSE for gamma if gamma_true is given
   \item bias_gamma: bias for gamma if gamma_true is given
   \item precision: precision for imputations 
   \item recall: recall for imputations 
}
}
\description{
Get parameter estimation and imputation results from a simulation
}
