% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getOmegas.R
\name{getOmegas}
\alias{getOmegas}
\title{Return omegas}
\usage{
getOmegas(k, omega)
}
\arguments{
\item{k}{category index}

\item{omega}{list of omegas}
}
\value{
vector of omegas for category k
}
\description{
Return omegas
}
