\name{pgmultinomr-package}
\alias{pgmultinomr-package}
\alias{pgmultinomr}
\docType{package}
\title{\packageTitle{pgmultinomr}}
\description{\packageDescription{pgmultinomr}}
\section{Package Content}{\packageIndices{pgmultinomr}}
\author{\packageAuthor{pgmultinomr}}
\section{Maintainer}{\packageMaintainer{pgmultinomr}}
\keyword{package}
