% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sim_dat.R
\name{sim_dat}
\alias{sim_dat}
\title{Simulate data to fit pgmultinom}
\usage{
sim_dat(
  n = 1000,
  miss_prob = 0,
  K = 6,
  p = 3,
  q = 5,
  covX = diag(3),
  allmiss = FALSE,
  null_scenario = FALSE,
  equal_probs = FALSE
)
}
\arguments{
\item{n}{sample size}

\item{miss_prob}{proportion of cases with missing outcome data}

\item{K}{number of outcome categories}

\item{p}{number of exposures}

\item{q}{number of covariates}

\item{covX}{covariance of exposure data for simulation}

\item{allmiss}{logical; if TRUE then all outcomes are missing for any case with missing outcome data, default is FALSE}

\item{null_scenario}{logical; if TRUE then exposures and covariates have no an effect on outcome, default is FALSE}

\item{equal_probs}{logical; if TRUE the simulation setting is equal probabilities, if FALSE the simulation setting is data probabilities}
}
\value{
summary of simulation results
}
\description{
Simulate data to fit pgmultinom
}
