% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sim_study.R
\name{sim_study}
\alias{sim_study}
\title{Run a simulation study using pgmultinom}
\usage{
sim_study(
  simnum = 1,
  niter = 1000,
  nburn = 500,
  n = 1000,
  miss_prob = 0,
  allmiss = FALSE,
  null_scenario = FALSE,
  equal_probs = FALSE
)
}
\arguments{
\item{simnum}{simulation number}

\item{niter}{number of MCMC iterations}

\item{nburn}{number of burn-in iterations}

\item{n}{sample size}

\item{miss_prob}{proportion of cases with missing outcome data}

\item{allmiss}{logical; if TRUE then all outcomes are missing for any case with missing outcome data, default is FALSE}

\item{null_scenario}{logical; if TRUE then exposures and covariates have no an effect on outcome, default is FALSE}

\item{equal_probs}{logical; if TRUE the simulation setting is equal probabilities, if FALSE the simulation setting is data probabilities}
}
\value{
a list with elements
}
\description{
Run a simulation study using pgmultinom
}
