% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pgmultinom.R
\name{pgmultinom}
\alias{pgmultinom}
\title{Bayesian categorical regression model with polya-gamma data augmentation}
\usage{
pgmultinom(niter, priors = NULL, y, x, w = NULL, intercept = TRUE)
}
\arguments{
\item{niter}{number of iterations}

\item{priors}{list of priors}

\item{y}{matrix of outcome data, with possible missing observations denoted by NA}

\item{x}{matrix of exposure data}

\item{w}{(optional) matrix of covariate data}

\item{intercept}{logical; if TRUE, include an intercept when fitting model}
}
\value{
a list with components
\itemize{
   \item y_original: matrix, categorical outcome data y for fitting 
   \item beta: list with components, the matrix of estimated exposure regression coefficients at each iteration
   \item beta.vec: list with components, the vectorized estimated exposure regression coefficients at each iteration
   \item gamma: list with components, the matrix of estimated covariate regression coefficients at each iteration
   \item gamma.vec: list with components, the vectorized estimated covariate regression coefficients at each iteration
   \item y.save: list with components, the matrix of outcomes with imputations at each iteration 
}
}
\description{
Bayesian categorical regression model with polya-gamma data augmentation
}
