% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pgmultinom.R
\name{pgmultinom}
\alias{pgmultinom}
\title{Fit Bayesian categorical regression model with polya-gamma data augmentation}
\usage{
pgmultinom(
  niter,
  priors = NULL,
  y,
  ycomplete = NULL,
  x,
  w = NULL,
  intercept = TRUE,
  beta_true = NULL,
  gamma_true = NULL
)
}
\arguments{
\item{niter}{number of iterations}

\item{priors}{list of priors}

\item{y}{matrix of outcome data, with possible missing observations denoted by NA}

\item{ycomplete}{matrix of complete outcome data, for validation}

\item{x}{matrix of exposure data}

\item{w}{(optional) matrix of covariate data}

\item{intercept}{logical; include an intercept?}

\item{beta_true}{(optional) list of true betas (exposure coefficients) for simulation study}

\item{gamma_true}{(optional) list of true gammas (covariate coefficients) for simulation study}
}
\value{
list of parameter estimates
}
\description{
Fit Bayesian categorical regression model with polya-gamma data augmentation
}
